requirements.txt
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.0
dnspython==2.8.0
exceptiongroup==1.3.0
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
ftfy==6.3.1
elasticsearch==8.15.1
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
open_clip_torch==3.2.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pymilvus==2.6.3
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
safetensors==0.6.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
Werkzeug==3.1.3
opencv-python
requests
